Added Better Auth Client & Remapped Server Routes
