+++ apps/client/src/lib/auth-client.ts
@@ -1,0 +1,4 @@
+import { createAuthClient } from "better-auth/react";
+export const authClient = createAuthClient({
+  baseURL: "http://localhost:3000/v1/api/auth",
+});
