+++ apps/server/package.json
@@ -14,6 +14,7 @@
 		"test": "vitest run --pass-with-no-tests"
 	},
 	"dependencies": {
+		"@fastify/autoload": "^6.3.1",
 		"@fastify/cors": "^11.0.1",
 		"@prisma/client": "6.7.0",
 		"better-auth": "^1.2.12",
