+++ apps/server/src/main/lib/auth.ts
@@ -1,7 +1,7 @@
 import { betterAuth } from "better-auth";
 import { prismaAdapter } from "better-auth/adapters/prisma";
 
-import prisma from "./prisma.js";
+import { prisma } from "./prisma.js";
 
 const trustedOrigin = process.env.CLIENT_ORIGIN || "http://localhost:3000";
 export const auth = betterAuth({
