+++ apps/server/src/main/lib/fastify.ts
@@ -1,0 +1,5 @@
+import Fastify from "fastify";
+
+export const fastify = Fastify({
+  logger: true,
+});
