+++ apps/server/src/main/routes/api/auth/auth.ts
@@ -1,0 +1,38 @@
+import type { FastifyInstance } from "fastify";
+
+import { auth } from "../../../lib/auth.js";
+
+export default async function (fastify: FastifyInstance) {
+  fastify.route({
+    method: ["GET", "POST"],
+    url: "/*",
+    async handler(request, reply) {
+      try {
+        const url = new URL(request.url, `http://${request.headers.host}`);
+
+        const headers = new Headers();
+        for (const [key, value] of Object.entries(request.headers)) {
+          if (value) headers.append(key, value.toString());
+        }
+
+        const req = new Request(url.toString(), {
+          method: request.method,
+          headers,
+          body: request.body ? JSON.stringify(request.body) : undefined,
+        });
+
+        const response = await auth.handler(req);
+
+        reply.status(response.status);
+        response.headers.forEach((value, key) => reply.header(key, value));
+        reply.send(response.body ? await response.text() : null);
+      } catch (error) {
+        fastify.log.error("Authentication Error:", error);
+        reply.status(500).send({
+          error: "Internal authentication error",
+          code: "AUTH_FAILURE",
+        });
+      }
+    },
+  });
+}
