+++ apps/server/src/main/routes/api/health/health.ts
@@ -1,0 +1,9 @@
+import type { FastifyInstance } from "fastify";
+
+export default async function (fastify: FastifyInstance) {
+  fastify.get("/", async (_, reply) => {
+    reply
+      .status(200)
+      .send({ status: "UP", timestamp: new Date().toISOString() });
+  });
+}
